requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
httpx==0.27.0
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic[email]==2.8.0   
python-multipart==0.0.6